+++ instrumentation/github.com/emicklei/go-restful/otelrestful/restful.go
@@ -61,6 +61,14 @@ func OTelFilter(service string, opts ...Option) restful.FilterFunction {
 			opts = append(opts, oteltrace.WithAttributes(rAttr))
 		}
 
+		if cfg.PublicEndpoint || (cfg.PublicEndpointFn != nil && cfg.PublicEndpointFn(r.WithContext(ctx))) {
+			opts = append(opts, oteltrace.WithNewRoot())
+			// Linking incoming span context if any for public endpoint.
+			if s := oteltrace.SpanContextFromContext(ctx); s.IsValid() && s.IsRemote() {
+				opts = append(opts, oteltrace.WithLinks(oteltrace.Link{SpanContext: s}))
+			}
+		}
+
 		ctx, span := tracer.Start(ctx, spanName, opts...)
 		defer span.End()